登录

+++ src/components/Menu/LeftMenu.vue
@@ -43,7 +43,6 @@
     },
     mounted() {
       this.getLeftRight()
-      this. queryPublicAreaUser()
     },
     methods: {
       getLeftRight() {
@@ -59,14 +58,15 @@
       },
       confirmLoginOut() {
         this.showTip = false
-        window.sessionStorage.clear() // 清楚session
-        window.localStorage.clear()
-        this.$router.replace('/login')
-      },
-      queryPublicAreaUser() {
-        this.$http('/UserController/queryPublicAreaUser').then(res => {
-
+        this.$http.post('/LoginController/loginOut').then(res => {
+          const data = res.data
+          if (data.code === 0) {
+            window.sessionStorage.clear() // 清楚session
+            window.localStorage.clear()
+            this.$router.replace('/login')
+          }
         })
+
       }
     },
   }
